Update README.md
This project aims to evaluate the creditworthiness of borrowers using a logistic regression model. The model is trained on a dataset of historical lending activity from a peer-to-peer lending services company. By analyzing various features of the loans, the model predicts whether a loan is healthy or poses a high risk of default.